+++ ecmascript/js_function.cpp
@@ -439,7 +439,7 @@ JSTaggedValue JSBoundFunction::ConstructInternal(EcmaRuntimeCallInfo *info)
     }
 
     JSHandle<TaggedArray> boundArgs(thread, func->GetBoundArguments());
-    const int32_t boundLength = boundArgs->GetLength();
+    const int32_t boundLength = static_cast<int32_t>(boundArgs->GetLength());
     const int32_t argsLength = info->GetArgsNumber() + boundLength;
     JSHandle<JSTaggedValue> undefined = thread->GlobalConstants()->GetHandledUndefined();
     EcmaRuntimeCallInfo *runtimeInfo =
